POC: add support for multiple zuul servers
Query both zuul.openstack.org and softwarefactory-project.io for
status of the jobs and builds a consolidated result table.

CHANGE IS NOT FULLY WORKING, FEEL FREE TO FIX IT.

Bugs:
- reloads entire page after number of seconds (not clear why)
- fails to perform correct AJAX query when current browser page
does not contain current changeset as it would use "0" insteaf of
correct number. The result is that Zuul will return nothing if
the changeset version is not the one currently being build. Need
to find alternative way to obtain it as the URL is not guaranteed to
have it.

Change-Id: Ifc89e456de88f77baa7ca772a852ad8e8d059d57